+++ EpikCord/__init__.py
@@ -8,16 +8,21 @@ import re
 from logging import getLogger
 from typing import *
 from urllib.parse import quote
-import io
-import os
 from .application import *
 from .channel import *
 from .client import *
 from .exceptions import *
 from .embed import *
-from .emojiemoji import *
+from .emoji import *
+from .file import *
+from .partials import *
+from .message import *
+from .utils import *
+from .intents import *
+from .interactions import *
+from .user import * 
+from .webhook import *
 
-T = TypeVar('T')
 logger = getLogger(__name__)
 __version__ = '0.4.12'
 
@@ -108,264 +113,10 @@ class Reaction:
         self.emoji: PartialEmoji = PartialEmoji(data.get("emoji"))
 
 
-class Message:
-    """Represents a Discord message.
-    
-    Attributes
-    ----------
-    id : str
-        The message ID.
-    channel_id : str
-        The channel ID the message was sent in.
-    author : :class:`User`
-        The author of the message 
-    guild_id: str
-        The Guild ID the message was sent in"""
-    def __init__(self, client, data: dict):
-        self.client = client
-        self.id: str = data.get("id")
-        self.channel_id: str = data.get("channel_id")
-        self.guild_id: Optional[str] = data.get("guild_id")
-        self.webhook_id: Optional[str] = data.get("webhook_id")
-        self.author: Optional[Union[WebhookUser, User]] = WebhookUser(data.get("author")) if data.get("webhook_id") else User(client, data.get("author"))
-        self.member: GuildMember = GuildMember(client, data.get("member")) if data.get("member") else None
-        # I forgot Message Intents are gonna stop this.
-        self.content: Optional[str] = data.get("content")
-        self.timestamp: str = data.get("timestamp")
-        self.edited_timestamp: Optional[str] = data.get("edited_timestamp")
-        self.tts: bool = data.get("tts")
-        self.mention_everyone: bool = data.get("mention_everyone")
-        self.mentions: Optional[List[MentionedUser]] = [MentionedUser(client, mention) for mention in data.get("mentions", [])]
-        self.mention_roles: Optional[List[int]] = data.get("mention_roles")
-        self.mention_channels: Optional[List[MentionedChannel]] = [MentionedChannel(channel) for channel in data.get("mention_channels", [])]
-        self.embeds: Optional[List[Embed]] = [Embed(**embed) for embed in data.get("embeds", [])]
-        self.reactions: Optional[List[Reaction]] = [Reaction(reaction) for reaction in data.get("reactions", [])]
-        self.nonce: Optional[Union[int, str]] = data.get("nonce")
-        self.pinned: bool = data.get("pinned")
-        self.type: int = data.get("type")
-        self.activity: Optional[MessageActivity] = MessageActivity(
-            data.get("activity")) if data.get("activity") else None
-        # Despite there being a PartialApplication, Discord don't specify what attributes it has
-        self.application: Application = Application(
-            data.get("application")) if data.get("application") else None
-        self.flags: int = data.get("flags")
-        self.referenced_message: Optional[Message] = Message(client, data.get("referenced_message")) if data.get("referenced_message") else None
-        self.interaction: Optional[MessageInteraction] = MessageInteraction(
-            client, data.get("interaction")) if data.get("interaction") else None
-        self.thread: Optional[Thread] = Thread(
-            data.get("thread")) if data.get("thread") else None
-
-        components: List[Any] = []
-        if data.get("components"):
-            for component in data.get("components"):
-                if component.get("type") == 1:
-                    components.append(MessageActionRow(component))
-                elif component.get("type") == 2:
-                    components.append(MessageButton(component))
-                elif components.get("type") == 3:
-                    components.append(MessageSelectMenu(component))
-                elif components.get("type") == 4:
-                    components.append(MessageTextInput(component))
-
-        self.components: Optional[List[Union[MessageTextInput, MessageSelectMenu, MessageButton]]] = components
-        self.stickers: Optional[List[StickerItem]] = [StickerItem(
-            sticker) for sticker in data.get("stickers", [])] or None
-
-    async def add_reaction(self, emoji: str):
-        emoji = quote(emoji)
-        logger.debug(f"Added a reaction to message ({self.id}).")
-        response = await self.client.http.put(f"channels/{self.channel_id}/messages/{self.id}/reactions/{emoji}/@me")
-        return await response.json()
-
-    async def remove_reaction(self, emoji: str, user=None):
-        emoji = quote(emoji)
-        logger.debug(
-            f"Removed reaction {emoji} from message ({self.id}) for user {user.username}.")
-        response = (
-            await self.client.http.delete(
-                f"channels/{self.channel_id}/messages/{self.id}/reactions/{emoji}/{user.id}"
-            )
-            if user
-            else await self.client.http.delete(
-                f"channels/{self.channel_id}/messages/{self.id}/reactions/{emoji}/@me"
-            )
-        )
-
-        return await response.json()
-
-    async def fetch_reactions(self, *, after, limit) -> List[Reaction]:
-        logger.debug(f"Fetching reactions from message ({self.id}).")
-        response = await self.client.http.get(f"channels/{self.channel_id}/messages/{self.id}/reactions?after={after}&limit={limit}")
-        return await response.json()
-
-    async def delete_all_reactions(self):
-        logger.debug(f"Deleting all reactions from message ({self.id}).")
-        response = await self.client.http.delete(f"channels/{self.channel_id}/messages/{self.id}/reactions")
-        return await response.json()
-
-    async def delete_reaction_for_emoji(self, emoji: str):
-        logger.debug(
-            f"Deleting a reaction from message ({self.id}) for emoji {emoji}.")
-        emoji = quote(emoji)
-        response = await self.client.http.delete(f"channels/{self.channel_id}/messages/{self.id}/reactions/{emoji}")
-        return await response.json()
-
-    async def edit(self, message_data: dict):
-        logger.debug(
-            f"Editing message {self.id} with message_data {message_data}.")
-        response = await self.client.http.patch(f"channels/{self.channel_id}/messages/{self.id}", data=message_data)
-        return await response.json()
-
-    async def delete(self):
-        logger.debug(f"Deleting message {self.id}.")
-        response = await self.client.http.delete(f"channels/{self.channel_id}/messages/{self.id}")
-        return await response.json()
-
-    async def pin(self, *, reason: Optional[str]):
-        headers = self.client.http.headers.copy()
-        if reason:
-            headers["X-Audit-Log-Reason"] = reason
-            logger.debug(f"Pinning message {self.id} with reason {reason}.")
-        else:
-            logger.debug(f"Pinning message {self.id}.")
-        response = await self.client.http.put(f"channels/{self.channel_id}/pins/{self.id}", headers=headers)
-        return await response.json()
-
-    async def unpin(self, *, reason: Optional[str]):
-        headers = self.client.http.headers.copy()
-        if reason:
-            headers["X-Audit-Log-Reason"] = reason
-            logger.debug(f"Unpinning message {self.id} with reason {reason}.")
-        else:
-            logger.debug(f"Unpinning message {self.id}.")
-        response = await self.client.http.delete(f"channels/{self.channel_id}/pins/{self.id}", headers=headers)
-        return await response.json()
-
-    async def start_thread(self, name: str, auto_archive_duration: Optional[int], rate_limit_per_user: Optional[int]):
-        logger.debug(
-            f"Starting thread for message {self.id} with name {name}, auto archive duration {auto_archive_duration or None}, ratelimit per user {rate_limit_per_user or None}.")
-        response = await self.client.http.post(f"channels/{self.channel_id}/messages/{self.id}/threads", data={"name": name, "auto_archive_duration": auto_archive_duration, "rate_limit_per_user": rate_limit_per_user})
-        # Cache it
-        self.client.guilds[self.guild_id].append(Thread(await response.json()))
-        return Thread(await response.json())
-
-    async def crosspost(self):
-        logger.debug(f"Crossposting message {self.id}.")
-        response = await self.client.http.post(f"channels/{self.channel_id}/messages/{self.id}/crosspost")
-        return await response.json()
-
-class File:
-    """
-    Represents a file. Sourced from Discord.py
-    """
-    def __init__(
-        self,
-        fp: Union[str, bytes, os.PathLike, io.BufferedIOBase],
-        filename: Optional[str] = None,
-        *,
-        spoiler: bool = False,
-    ):
-        if isinstance(fp, io.IOBase):
-            if not (fp.seekable() and fp.readable()):
-                raise ValueError(f'File buffer {fp!r} must be seekable and readable')
-            self.fp = fp
-            self._original_pos = fp.tell()
-        else:
-                self.fp = open(fp, 'rb')
-                self._original_pos = 0
-        self._closer = self.fp.close
-        self.fp.close = lambda: None
-        
-        if filename is None:
-            if isinstance(fp, str):
-                _, self.filename = os.path.split(fp)
-            else:
-                self.filename = getattr(fp, 'name', None)
-        else:
-            self.filename = filename
-        if spoiler and self.filename is not None and not self.filename.startswith('SPOILER_'):
-            self.filename = 'SPOILER_' + self.filename
-
-            self.spoiler = spoiler or (self.filename is not None and self.filename.startswith('SPOILER_'))
-
-        def reset(self, *, seek: Union[int, bool] = True) -> None:
-            if seek:
-                self.fp.seek(self._original_pos)
-
-        def close(self) -> None:
-            self.fp.close = self._closer
-            self._closer()
-
-class Messageable:
-    def __init__(self, client, channel_id: str):
-        self.id: str = channel_id
-        self.client = client
-
-    async def fetch_messages(self, *, around: Optional[str] = None, before: Optional[str] = None, after: Optional[str] = None, limit: Optional[int] = None) -> List[Message]:
-        response = await self.client.http.get(f"channels/{self.id}/messages", params={"around": around, "before": before, "after": after, "limit": limit})
-        data = await response.json()
-        return [Message(self.client, message) for message in data]
-
-    async def fetch_message(self, *, message_id: str) -> Message:
-        response = await self.client.http.get(f"channels/{self.id}/messages/{message_id}")
-        data = await response.json()
-        return Message(self.client, data)
-
-    async def send(self, content: Optional[str] = None, *, embeds: Optional[List[dict]] = None, components=None, tts: Optional[bool] = False, allowed_mentions=None, sticker_ids: Optional[List[str]] = None, attachments: List[File]=None, suppress_embeds: bool = False) -> Message:
-        payload = {}
-
-        if content:
-            payload["content"] = content
-
-        if embeds:
-            payload["embeds"] = [embed.to_dict() for embed in embeds]
-
-        if components:
-            payload["components"] = [component.to_dict()
-                                     for component in components]
-
-        if tts:
-            payload["tts"] = tts
-
-        if allowed_mentions:
-            payload["allowed_mentions"] = allowed_mentions.to_dict()
-
-        if sticker_ids:
-            payload["sticker_ids"] = sticker_ids
-
-        if attachments:
-            payload["attachments"] = [attachment.to_dict()
-                                      for attachment in attachments]
-
-        if suppress_embeds:
-            payload["suppress_embeds"] = 1 << 2
-
-        response = await self.client.http.post(f"channels/{self.id}/messages", json=payload)
-        data = await response.json()
-        return Message(self.client, data)
 
 
-class User(Messageable):
-    def __init__(self, client, data: dict):
-        super().__init__(client, data["id"])
-        self.data = data
-        self.client = client
-        self.id: str = data.get("id")
-        self.username: str = data.get("username")
-        self.discriminator: str = data.get("discriminator")
-        self.avatar: Optional[str] = data.get("avatar")
-        self.bot: Optional[bool] = data.get("bot")
-        self.system: Optional[bool] = data.get("system")
-        self.mfa_enabled: bool = data.get("mfa_enabled")
-        self.banner: Optional[str] = data.get("banner")
-        # the user's banner color encoded as an integer representation of hexadecimal color code
-        self.accent_color: Optional[int] = data.get("accent_color")
-        self.locale: Optional[str] = data.get("locale")
-        self.verified: bool = data.get("verified")
-        self.email: Optional[str] = data.get("email")
-        self.flags: int = data.get("flags")
-        self.premium_type: int = data.get("premium_type")
-        self.public_flags: int = data.get("public_flags")
+
+
 class EventHandler:
     # Class that'll contain all methods that'll be called when an event is triggered.
 
@@ -2183,34 +1934,7 @@ class Guild:
 
         return _figure_out_channel_type(await self.client.http.post(f"/guilds/{self.id}/channels", json=data, headers=headers))
 
-class WebhookUser:
-    def __init__(self, data: dict):
-        self.webhook_id: str = data.get("webhook_id")
-        self.username: str = data.get("username")
-        self.avatar: str = data.get("avatar")
 
-
-class Webhook:
-    def __init__(self, client, data: dict = None):
-        """
-        Don't pass in data if you're making a webhook, the lib passes data to construct an already existing webhook
-        """
-        self.client = client
-        self.data = data
-        if data:
-            self.id: str = data.get("id")
-            self.type: str = "Incoming" if data.get(
-                "type") == 1 else "Channel Follower" if data.get("type") == 2 else "Application"
-            self.guild_id: Optional[str] = data.get("guild_id")
-            self.channel_id: Optional[str] = data.get("channel_id")
-            self.user: Optional[User] = User(client, data.get("user"))
-            self.name: Optional[str] = data.get("name")
-            self.avatar: Optional[str] = data.get("avatar")
-            self.token: Optional[str] = data.get("token")
-            self.application_id: Optional[str] = data.get("application_id")
-            self.source_guild: Optional[PartialGuild] = PartialGuild(
-                data.get("source_guild"))
-            self.url: Optional[str] = data.get("url")
 
 class Modal:
     def __init__(self, *, title: str, custom_id: str, components: List[MessageActionRow]):
@@ -2225,224 +1949,7 @@ class Modal:
             "components": self.components
         }
 
-class BaseInteraction:
-    def __init__(self, client, data: dict):
-        self.id: str = data.get("id")
-        self.client = client
-        self.type: int = data.get("type")
-        self.application_id: int = data.get("application_id")
-        self.data: dict = data
-        self.interaction_data: Optional[dict] = data.get("data")
-        self.guild_id: Optional[str] = data.get("guild_id")
-        self.channel_id: Optional[str] = data.get("channel_id")
-        self.member: Optional[GuildMember] = GuildMember(client, data.get("member")) if data.get("member") else None
-        self.user: Optional[User] = User(client, data.get("user")) if data.get("user") else None
-        self.token: str = data.get("token")
-        self.version: int = data.get("version")
-        self.locale: Optional[str] = data.get("locale")
-        self.guild_locale: Optional[str] = data.get("guild_locale")
-        self.original_response: Optional[Message] = None # Can't be set on construction.
-        self.followup_response: Optional[Message] = None # Can't be set on construction.
 
-    def is_application_command(self):
-        return self.type == 2
-
-    def is_message_component(self):
-        return self.type == 3
-
-    def is_autocomplete(self):
-        return self.type == 4
-    
-    def is_modal_submit(self):
-        return self.type == 5
-    async def fetch_original_response(self, *, skip_cache: Optional[bool] = False):
-        if not skip_cache and self.original_response:
-            return self.original_response
-        message_data = await self.client.http.get(f"/webhooks/{self.application_id}/{self.token}/messages/@original")
-        self.original_response: Message = Message(self.client, message_data)
-        return self.original_response
-    
-    async def edit_original_response(self, *, tts: bool = False, content: Optional[str] = None, embeds: Optional[List[Embed]] = None, allowed_mentions = None, components: Optional[List[Union[MessageButton, MessageSelectMenu, MessageTextInput]]] = None, attachments: Optional[List[Attachment]] = None, suppress_embeds: Optional[bool] = False, ephemeral: Optional[bool] = False) -> None:
-
-        message_data = {
-            "tts": tts,
-            "flags": 0
-        }
-
-        if suppress_embeds:
-            message_data["flags"] += 1 << 2
-        if ephemeral:
-            message_data["flags"] += 1 << 6
-
-        if content:
-            message_data["content"] = content
-        if embeds:
-            message_data["embeds"] = [embed.to_dict() for embed in embeds]
-        if allowed_mentions:
-            message_data["allowed_mentions"] = allowed_mentions.to_dict()
-        if components:
-            message_data["components"] = [component.to_dict() for component in components]
-        if attachments:
-            message_data["attachments"] = [attachment.to_dict() for attachment in attachments]
-
-        new_message_data = await self.client.http.patch(f"/webhooks/{self.application_id}/{self.token}/messages/@original", json = message_data)
-        self.original_response: Message = Message(self.client, new_message_data)
-        return self.original_response
-    
-    async def delete_original_response(self):
-        await self.client.http.delete(f"/webhooks/{self.application_id}/{self.token}/messages/@original")
-
-    async def create_followup(self, *, tts: bool = False, content: Optional[str] = None, embeds: Optional[List[Embed]] = None, allowed_mentions = None, components: Optional[List[Union[MessageButton, MessageSelectMenu, MessageTextInput]]] = None, attachments: Optional[List[Attachment]] = None, suppress_embeds: Optional[bool] = False, ephemeral: Optional[bool] = False) -> None:
-
-        message_data = {
-            "tts": tts,
-            "flags": 0
-        }
-
-        if suppress_embeds:
-            message_data["flags"] += 1 << 2
-        if ephemeral:
-            message_data["flags"] += 1 << 6
-
-        if content:
-            message_data["content"] = content
-        if embeds:
-            message_data["embeds"] = [embed.to_dict() for embed in embeds]
-        if allowed_mentions:
-            message_data["allowed_mentions"] = allowed_mentions.to_dict()
-        if components:
-            message_data["components"] = [component.to_dict() for component in components]
-        if attachments:
-            message_data["attachments"] = [attachment.to_dict() for attachment in attachments]
-
-        new_message_data = await self.client.http.post(f"/webhooks/{self.application_id}/{self.token}/", json = message_data)
-        self.followup_response: Message = Message(self.client, new_message_data)
-        return self.followup_response
-
-    async def edit_followup(self, *, tts: bool = False, content: Optional[str] = None, embeds: Optional[List[Embed]] = None, allowed_mentions = None, components: Optional[List[Union[MessageButton, MessageSelectMenu, MessageTextInput]]] = None, attachments: Optional[List[Attachment]] = None, suppress_embeds: Optional[bool] = False, ephemeral: Optional[bool] = False) -> None:
-
-        message_data = {
-            "tts": tts,
-            "flags": 0
-        }
-
-        if suppress_embeds:
-            message_data["flags"] += 1 << 2
-        if ephemeral:
-            message_data["flags"] += 1 << 6
-
-        if content:
-            message_data["content"] = content
-        if embeds:
-            message_data["embeds"] = [embed.to_dict() for embed in embeds]
-        if allowed_mentions:
-            message_data["allowed_mentions"] = allowed_mentions.to_dict()
-        if components:
-            message_data["components"] = [component.to_dict() for component in components]
-        if attachments:
-            message_data["attachments"] = [attachment.to_dict() for attachment in attachments]
-
-        await self.client.http.patch(f"/webhook/{self.application_id}/{self.token}/", json = message_data)
-
-    async def delete_followup(self):
-        return await self.client.http.delete(f"/webhook/{self.application_id}/{self.token}/")
-
-class MessageComponentInteraction(BaseInteraction):
-    def __init__(self, client, data: dict):
-        super().__init__(client, data)
-        self.custom_id: str = self.data.get("custom_id")
-        self.component_type: Optional[int] = self.data.get("component_type")
-        self.values: Optional[dict] = [MessageSelectMenuOption(option) for option in self.data.get("values", [])]
-
-    async def update(self, *, tts: bool = False, content: Optional[str] = None, embeds: Optional[List[Embed]] = None, allowed_mentions = None, components: Optional[List[Union[MessageButton, MessageSelectMenu, MessageTextInput]]] = None, attachments: Optional[List[Attachment]] = None, suppress_embeds: Optional[bool] = False, ephemeral: Optional[bool] = False) -> None:
-
-        message_data = {
-            "tts": tts,
-            "flags": 0
-        }
-
-        if suppress_embeds:
-            message_data["flags"] += 1 << 2
-        if ephemeral:
-            message_data["flags"] += 1 << 6
-
-        if content:
-            message_data["content"] = content
-        if embeds:
-            message_data["embeds"] = [embed.to_dict() for embed in embeds]
-        if allowed_mentions:
-            message_data["allowed_mentions"] = allowed_mentions.to_dict()
-        if components:
-            message_data["components"] = [component.to_dict() for component in components]
-        if attachments:
-            message_data["attachments"] = [attachment.to_dict() for attachment in attachments]
-
-        payload = {
-            "type": 7,
-            "data": message_data
-        }
-
-        await self.client.http.patch(f"/interaction/{self.id}/{self.token}/callback", json = payload)
-
-    async def defer_update(self):
-        await self.client.http.post(f"/interaction/{self.id}/{self.token}/callback", json = {
-            "type": 6
-        })
-
-
-class ModalSubmitInteraction(BaseInteraction):
-    def __init__(self, client, data: dict):
-        super().__init__(client, data)
-        self.components: List[Union[MessageButton, MessageSelectMenu, MessageTextInput]] = []
-        for component in data.get("components"):
-            if component.get("type") == 2:
-                self.components.append(MessageButton(component))
-            elif component.get("type") == 3:
-                self.components.append(MessageSelectMenu(component))
-            elif component.get("type") == 4:
-                self.components.append(MessageTextInput(component))
-
-
-
-class AutoCompleteInteraction(BaseInteraction):
-    def __init__(self, client, data: dict):
-        super().__init__(client, data)
-        self.options: List[ApplicationCommandOption] = [ApplicationCommandOption(option) for option in data.get("options", [])]
-
-    async def reply(self, choices: List[SlashCommandOptionChoice]) -> None:
-        payload = {
-            "type": 9,
-            "data": []
-        }
-
-        for choice in choices:
-            if not isinstance(choice, SlashCommandOptionChoice):
-                raise TypeError(f"{choice} must be of type SlashCommandOptionChoice")
-            payload["data"]["choices"].append(choice.to_dict())
-
-        await self.client.http.post(f"/interactions/{self.id}/{self.token}/callback", json = payload)
-
-
-
-class ResolvedDataHandler:
-    def __init__(self, client, resolved_data: dict):
-        self.data: dict = resolved_data # In case we miss anything and people can just do it themselves
-        self.users: dict = [User(client, user) for user in self.data.get("users", [])]
-
-        self.members: dict = [GuildMember()]
-        self.roles: dict = self.data["roles"]
-        self.channels: dict = self.data["channels"]
-
-
-    
-    
-class UserCommandInteraction(ApplicationCommandInteraction):
-    def __init__(self, client, data: dict):
-        super().__init__(client, data)
-        self.target_id: str = data.get("target_id")
-
-class MessageCommandInteraction(UserCommandInteraction):
-    ... # Literally the same thing.
 
 class Invite:
     def __init__(self, data: dict):
@@ -2500,10 +2007,7 @@ class MentionedUser(User):
         super().__init__(client, data)
         self.member = GuildMember(client, data.get("member")) if data.get("member") else None
 
-class MessageActivity:
-    def __init__(self, data: dict):
-        self.type: int = data.get("type")
-        self.party_id: Optional[str] = data.get("party_id")
+
 
 
 class AllowedMention:
@@ -2525,22 +2029,6 @@ class MessageInteraction:
         self.member: Optional[GuildMember] = GuildMember(client, data.get("member")) if data.get("member") else None
         self.user: User = User(client, data.get("user"))
 
-class PartialUser:
-    def __init__(self, data: dict):
-        self.data: dict = data
-        self.id: str = data.get("id")
-        self.username: str = data.get("username")
-        self.discriminator: str = data.get("discriminator")
-        self.avatar: Optional[str] = data.get("avatar")
-
-
-class PartialGuild:
-    def __init__(self, data):
-        self.data: dict = data
-        self.id: str = data.get("id")
-        self.name: str = data.get("name")
-        self.permissions: int = int(data.get("permissions"))
-        self.features: List[str] = data.get("features")
 
 
 class SlashCommand(ApplicationCommand):
@@ -2652,117 +2140,7 @@ class Webhook:  # Not used for making webhooks.
         self.source_channel: Optional[SourceChannel] = SourceChannel(data.get("source_channel"))
         self.url: Optional[str] = data.get("url")
 
-class Intents:
-    def __init__(self, *, intents: Optional[int] = None):
-        self.value = intents or 0
 
-    @property
-    def guilds(self):
-        self.value += 1 << 0
-        return self
-
-    @property
-    def guild_members(self):
-        self.value += 1 << 1
-        return self
-
-    @property
-    def guild_bans(self):
-        self.value += 1 << 2
-        return self
-
-    @property
-    def guild_emojis_and_stickers(self):
-        self.value += 1 << 3
-        return self
-
-    @property
-    def guild_integrations(self):
-        self.value += 1 << 4
-        return self
-
-    @property
-    def guild_webhooks(self):
-        self.value += 1 << 5
-        return self
-
-    @property
-    def guild_invites(self):
-        self.value += 1 << 6
-        return self
-
-    @property
-    def guild_voice_states(self):
-        self.value += 1 << 7
-        return self
-
-    @property
-    def guild_presences(self):
-        self.value += 1 << 8
-        return self
-
-    @property
-    def guild_messages(self):
-        self.value += 1 << 9
-        return self
-
-    @property
-    def guild_message_reactions(self):
-        self.value += 1 << 10
-        return self
-
-    @property
-    def guild_message_typing(self):
-        self.value += 1 << 11
-        return self
-
-    @property
-    def direct_messages(self):
-        self.value += 1 << 12
-        return self
-
-    @property
-    def direct_message_reactions(self):
-        self.value += 1 << 13
-        return self
-
-    @property
-    def direct_message_typing(self):
-        self.value += 1 << 14
-        return self
-
-    @property
-    def all(self):
-        for attr in dir(self):
-            if attr not in ["value", "all", "none", "remove_value", "add_intent"]:
-                getattr(self, attr)
-        return self
-
-    @property
-    def none(self):
-        self.value = 0
-
-    def remove_intent(self, intent: str) -> int:
-        try:
-            attr = getattr(self, intent.lower())
-        except AttributeError:
-            raise InvalidIntents(
-                f"Intent {intent.lower()} is not a valid intent.")
-        self.value -= attr.value
-        return self.value
-
-    def add_intent(self, intent: str) -> int:
-        try:
-            attr = getattr(self, intent)
-        except AttributeError:
-            raise InvalidIntents(f"Intent {intent} is not a valid intent.")
-        self.value += attr
-        return self.value
-
-    @property
-    def message_content(self):
-        self.value += 1 << 15
-        return self
 
     # TODO: Add some presets such as "Moderation", "Logging" etc.
 
@@ -3010,122 +2388,3 @@ class Paginator:
 
     def current(self) -> Embed:
         return self.pages[self.current_index] 
-
-class Utils:
-    """
-    A utility class, used to make difficult things easy.
-
-    Attributes:
-    ----------
-    client: Client
-        The client that this utility class is attached to.
-        
-    """
-
-    def __init__(self, client):
-        self.client = client
-        self. _MARKDOWN_ESCAPE_SUBREGEX = '|'.join(
-            r'\{0}(?=([\s\S]*((?<!\{0})\{0})))'.format(c) for c in ('*', '`', '_', '~', '|'))
-
-        self._MARKDOWN_ESCAPE_COMMON = r'^>(?:>>)?\s|\[.+\]\(.+\)'
-
-        self._MARKDOWN_ESCAPE_REGEX = re.compile(
-            fr'(?P<markdown>{self._MARKDOWN_ESCAPE_SUBREGEX}|{self._MARKDOWN_ESCAPE_COMMON})', re.MULTILINE)
-
-        self._URL_REGEX = r'(?P<url><[^: >]+:\/[^ >]+>|(?:https?|steam):\/\/[^\s<]+[^<.,:;\"\'\]\s])'
-
-        self._MARKDOWN_STOCK_REGEX = fr'(?P<markdown>[_\\~|\*`]|{self._MARKDOWN_ESCAPE_COMMON})'
-
-
-    def channel_from_type(self, channel_data: dict):
-        channel_type = channel_data.get("type")
-        if channel_type == 0:
-            return GuildTextChannel(self.client, channel_data)
-        elif channel_type == 1:
-            return DMChannel(self.client, channel_data)
-        elif channel_type == 2:
-            return VoiceChannel(self.client, channel_data)
-        elif channel_type == 4:
-            return ChannelCategory(self.client, channel_data)
-        elif channel_type == 5:
-            return GuildNewsChannel(self.client, channel_data)
-        elif channel_type == 6:
-            return GuildStoreChannel(self.client, channel_data)
-        elif channel_type == 10:
-            return GuildNewsThread(self.client, channel_data)
-        elif channel_type == 11:
-            return Thread(self.client, channel_data)
-        elif channel_type == 12:
-            return PrivateThread(self.client, channel_data)
-        elif channel_type == 13:
-            return GuildStageChannel(self.client, channel_data)
-
-    def compute_timedelta(self, dt: datetime.datetime):
-        if dt.tzinfo is None:
-            dt = dt.astimezone()
-        now = datetime.datetime.now(datetime.timezone.utc)
-        return max((dt - now).total_seconds(), 0)
-
-
-    async def sleep_until(self, when: Union[datetime.datetime, int, float], result: Optional[T] = None) -> Optional[T]:
-        if when == datetime.datetime:
-            delta = self.compute_timedelta(when)
-
-        return await asyncio.sleep(delta if when == datetime.datetime else when, result)
-
-
-    def remove_markdown(self, text: str, *, ignore_links: bool = True) -> str:
-        def replacement(match):
-            groupdict = match.groupdict()
-            return groupdict.get('url', '')
-
-        regex = self._MARKDOWN_STOCK_REGEX
-        if ignore_links:
-            regex = f'(?:{self._URL_REGEX}|{regex})'
-        return re.sub(regex, replacement, text, 0, re.MULTILINE)
-
-
-    def escape_markdown(self, text: str, *, as_needed: bool = False, ignore_links: bool = True) -> str:
-        if not as_needed:
-
-            def replacement(match):
-                groupdict = match.groupdict()
-                if is_url := groupdict.get('url'):
-                    return is_url
-                return '\\' + groupdict['markdown']
-
-            regex = self._MARKDOWN_STOCK_REGEX
-            if ignore_links:
-                regex = f'(?:{self._URL_REGEX}|{regex})'
-            return re.sub(regex, replacement, text, 0, re.MULTILINE)
-        else:
-            text = re.sub(r'\\', r'\\\\', text)
-            return self._MARKDOWN_ESCAPE_REGEX.sub(r'\\\1', text)
-
-
-    def escape_mentions(self, text: str) -> str:
-        return re.sub(r'@(everyone|here|[!&]?[0-9]{17,20})', '@\u200b\\1', text)
-
-
-    def utcnow(self) -> datetime.datetime:
-        return datetime.datetime.now(datetime.timezone.utc)
-
-    def cancel_tasks(self, loop) -> None:
-        tasks = {t for t in asyncio.all_tasks(loop=loop) if not t.done()}
-
-        if not tasks:
-            return
-
-        for task in tasks:
-            task.cancel()
-        logger.debug(f"Cancelled {len(tasks)} tasks")
-        loop.run_until_complete(asyncio.gather(*tasks, return_exceptions=True))
-
-
-    def cleanup_loop(self, loop) -> None:
-        try:
-            self.cancel_tasks(loop)
-            logger.debug("Shutting down async generators.")
-            loop.run_until_complete(loop.shutdown_asyncgens())
-        finally:
-            loop.close()
