+++ EpikCord/application.py
@@ -1,5 +1,9 @@
 from typing import (Optional, Union, List,)
 from .exceptions import InvalidArgumentType
+from .partials import PartialUser
+from .embed import Embed
+from .interactions import *
+
 #from .__init__ import PartialUser, Team, BaseInteraction, ResolvedDataHandler, Modal,InvalidArgumentType,Embed,MessageButton,MessageSelectMenu,MessageTextInput,Attachment
 #temp import (So many!)
 # we have to make sure there is no circular imports, someone help us!
